Add dummy stone plating parts to make it easier to make casts
Defines a new item class, DummyMaterialItem, to create the "bases" for casts for things not made in the part builder
Also defines a new tag to allow adding items to the part chest that are not tool parts
The rest is just normal item stuff, taking advantage of the fact part builders can produce non-tool parts

Just so its written down somewhere, the reason I did not just use copper here is because I want to require casting to make plating, as opposed to making plate armor in the part builder via copper. Keep the melter gate for the mechanic.