Create loot injectors
Loot injectors are a JSON controlled method of adding new loot to an existing loot table. Allows JSON addons to easily inject loot without conflict, and may be useful for a modpack maker wanting to make a smaller change
Will eventually want to move this to Mantle, but will do that in 1.20.
System planned to be used for ancient tools soon