Revamp frying into smelting and make it craftable
Was originally planned for the frying pan, but the frying pan got revamped so no longer needs it. Instead making it an upgrade
As an upgrade, it works on mining blocks, damaging mobs, shooting bows, taking damage in armor, etc. and uses any smelting recipe
Also includes a mined block count in the tool after harvest hook now