Implement the melting pan
Yeah, only been referenced in the last 5 commits or so
Melting pan mines meltable blocks and stores the liquid in itself.
Cannot mine if it cannot melt, meaning the result fluid needs to match the pan fluid and it cannot be full.
Melting pan is registered as a staff, so you can make a weird mining staff or just ignore the mining entirely in favor of a fluid staff build.