Fix longspam from checking for texture existence
Turns out the new sprite getter includes a logger if a sprite doesn't exist. So have to bring back our old system to check for texture existence.
Using the first method (folder scan) since it was most efficient and early load is early enough for it (since texture lookup is now in the late stage).