Add a new broad binding and turn round plates into adze heads
A large plate is half a large head, but a round plate is not half a small head, so it felt weird. Redesigning it into an adze head makes it more consistent and better describes its usages.
The adze side of the pickadze is now a bit more obvious to reduce the number of people mixing it up with a pickaxe
Excavators replaces their second large plate with a tough binding
Vein hammers replaced their pickaxe head with a tough binding
Scythes and broad axes now use a tough binding instead of a tool binding