Use the smithing template for the netherite modifier
For symmetry, diamond and emerald also use a "template", former being a gold cast and latter an obsidian pane. Opens up some freedom to reuse those items as materials (not that we plan to do so)
Due to the higher cost of netherite, it no longer requires a gem to apply (gets the gem from the template). Means its 1 less upgrade slot to apply unbreakable. Will see what the consequences of that are and if we like it.